feat: move biz logic from widget to cubit
